fix : 역전 검사를 판정 앞으로 · 사용 축 상한을 V5 에 실제로 건다
리뷰 3회에서 나온 결함을 고친다.

BLOCKER — 역전 검사가 막으려던 경우에만 안 돌았다
  statsAggregateStep 은 dataset != CLEAN 과 verdict != PASS 에서 먼저 반환한다.
  역전이 ILLEGAL_TRANSITION 오탐을 내면 검출이 1건 이상이라 verdict 가 FAIL 이 되고,
  그 조기 반환에 걸려 검사가 한 줄도 안 돌았다. 게다가 판정은 앞 Step 이 이미
  커밋해서, 거기서 죽어도 증적에는 verdict=PASS 가 남았다.
  → assertFrozenStep 으로 옮긴다. 판정 앞이고, CORRUPT 도 지나고, 죽으면 verdict 가
    NULL 로 남아 "판정이 안 났다" 가 그대로 신호가 된다.
  → 옮겨도 CORRUPT 288건 게이트가 안 깨지는 것을 실측으로 확인했다
    (coupon_clean · coupon_corrupt · coupon_bench 모두 역전 0쌍).

에러 코드를 가른다 — VERIFICATION-023 HISTORY_ORDER_INVERTED
  역전은 이미 저장된 행의 속성이라 쓰기를 멈춰도 안 사라진다.
  DATASET_MUTATED_DURING_RUN 은 "멈추고 재실행" 계열이라 운영자를 무한 재시도로
  보낸다. docs/11 이 세운 재시도 가능성 갈림 규칙을 따른다.

가드 메시지를 사실로 만든다
  "V5 는 얼린 상한까지 접은 값을 읽었다" 가 거짓이었다 — APPLY_USAGE_COUNTS 에
  id 상한이 없었다. 상한을 실제로 걸어 이력 축과 대칭으로 만든다.
  상한은 기본값 없이 꺼낸다(requireFrozenMaxUsage) — 0 이 들어가면 사용을 한 건도
  안 세어 전 발급건이 오탐이 되는, 조용히 틀리는 방향이다.

latestUsageId 의 used_at 술어를 뺀다
  "PK 범위 조회라 비용이 없다" 가 거짓이었다 — used_at 인덱스가 없어 132만 행
  전수 스캔이다(실측 type=ALL · rows=1,313,897 · 0.32초). id 가 오토인크리먼트라
  술어 없이도 가드의 뜻이 같다. 실측 0.32초 → 0.0025ms.

문서 수치를 고친다 — EXPLAIN 추정치를 실측처럼 옮겼다
  17.5만 그룹 → 실측 105만 (EXPLAIN 의 rows= 는 6배 어긋난다)
  asof_state 실행당 770 MiB · 6배 → 770 은 4실행 합계. 실행당 192 MiB · 1.5배
  행당 3.2ms → 46.5초 ÷ 105만 = 약 44µs
  "인덱스로는 줄일 수 없는 축" → 안 쟀다. 집계 인덱스가 필터 컬럼을 안 덮어
  커버링 후보가 남아 있다. docs/13 인덱스 실측으로 넘긴다.
  옛 17초 데이터포인트는 지우지 않고 "화해 못 함" 으로 남긴다.
  docs/13 · docs/17 의 낡은 문장도 그 자리에서 고친다.

테스트 — 이 결함들이 통과한 이유가 여기다
  잡 수준 2건: 역전이 판정 전에 죽는지 · 사용 축 상한이 실제로 실리는지
  창 축 3건: b 가 상한 위여도 세는지 · a 가 상한 위면 안 세는지 · asOf 밖
  돌연변이로 확인했다 — b 에 창을 걸면, 판정 앞 검사를 빼면 각각 빨개진다.

기타
  .alerts.json 제거 — PPT 용 임시 덤프가 커밋에 딸려 들어갔다. 참조 0건에
  알림 정의 45개가 통째로 복제돼 rules yml 과 갈릴 자리였다.
  가로챈 javadoc 원복 · 죽은 null 검사 제거 · 쌍 카운트 int → long
  얼림 축 개수(넷 → 다섯)와 assertStillFrozen 의 비대칭 근거를 문서에 맞춘다

+++ core/src/main/java/com/kafkick/core/verification/StatsRepository.java
@@ -111,9 +111,14 @@ public interface StatsRepository {
      * 원인을 그 자리에 세운다 — 그래야 오탐이 <i>"검증기 버그"</i> 가 아니라
      * <i>"이력 시각이 역전됐다"</i> 로 보고된다.
      *
-     * <p>창은 리플레이와 같다 — {@code id <= maxHistoryId AND created_at <= asOf}.
+     * <p><b>창은 {@code a} 쪽에만 리플레이와 같다</b> — {@code a.id <= maxHistoryId AND
+     * a.created_at <= asOf}. {@code b} 는 시각으로 이미 갇힌다({@code b.created_at <
+     * a.created_at}) 므로 따로 안 자른다. 비용 근거는 {@code StatsJdbcAdapter} 의
+     * {@code COUNT_OUT_OF_ORDER} 에 적었다.
+     *
+     * <p>세는 것이 행이 아니라 <b>쌍</b>이라 {@code long} 이다 — 한 발급건 안에서 제곱으로 는다.
      */
-    int countOutOfOrderHistoryPairs(LocalDateTime asOf, long frozenMaxHistoryId);
+    long countOutOfOrderHistoryPairs(LocalDateTime asOf, long frozenMaxHistoryId);
 
     /** 위 질의가 0 이 아닐 때 메시지에 실을 발급건 id 표본. */
     List<Long> sampleIssuancesWithBrokenIssueHistory(
